Fix unbound scheduling of trim work items
This condition should almost never be hit except if .Cache(limit) is ran in a loop, then it is possible that a state is achieved where the amount of scheduled trim operations will be equal to the amount of .Cache() calls, which can be a lot.

To be honest, CacheManager, QuickList and CacheStaticHolder need to be re-organized into something that looks less offensive, doesn't use circular static dependencies and lends itself into easier testability but I CBA to do it now.